+++ src/file.ts
@@ -1,7 +1,6 @@
 import { mkdir, readFile, readdir, writeFile } from "node:fs/promises";
 import path from "node:path";
 import prettier from "prettier";
-import type { z } from "zod";
 
 import { generateKeys } from "./generate-translation-keys";
 import {
@@ -14,39 +13,7 @@ import {
 } from "./state";
 import { translate } from "./translate";
 import { TranslationSchema } from "./types";
-
-const writeJSON = async (
-	prettierConfig: prettier.Options,
-	folder: string,
-	fileName: string,
-	content: Record<string, unknown> | unknown[],
-): Promise<void> => {
-	const pathToFile = path.resolve(folder, fileName);
-	const fileContent = JSON.stringify(content, null, 2);
-	const filePrettified = await prettier.format(fileContent, {
-		...prettierConfig,
-		parser: "json",
-	});
-
-	await mkdir(folder, { recursive: true });
-	await writeFile(pathToFile, filePrettified, "utf-8");
-};
-
-const parseJSON = <T extends z.ZodTypeAny>(
-	type: T,
-	input: string,
-): z.output<T> => {
-	return type.parse(JSON.parse(input));
-};
-
-export const readJSON = async <T extends z.ZodTypeAny>(
-	type: T,
-	path: string,
-): Promise<z.output<T>> => {
-	const content = await readFile(path, "utf-8");
-
-	return parseJSON(type, content);
-};
+import { readJSON, writeJSON } from "./utils";
 
 export async function initializeState({
 	translationsPath,
